+++ HISTORY.md
@@ -1,4 +1,8 @@
 # History
 
+## v1.1.0 2018 July 19
+- Added `configure` and `interval` methods, as well as the `active` computed property
+- Updates with empty title text are correctly supported (they are used to reset the terminal title)
+
 ## v1.0.0 2018 July 19
 - Initial working release